// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/ot/shared/dto/post-ot-mst-checklist.dto.ts
export class PostOTMSTCheckList_DTO {
    CheckListId: number = 0;
    ServiceItemId: number = null;
    CheckListName: string = null;
    DisplayName: string = null;
    InputType: string = null;
    IsMandatory: boolean = false;
    DisplaySequence: number = null;
    LookUp: string = null;
    IsActive: boolean = false;
}